🚚 move functions to make URLs absolute in `serializationUtils`
