+++ packages/rum-recorder/src/domain/record/serialize.ts
@@ -2,7 +2,14 @@ import { display } from '@datadog/browser-core'
 import { PRIVACY_ATTR_NAME, PRIVACY_ATTR_VALUE_HIDDEN } from '../../constants'
 import { nodeShouldBeHidden } from './privacy'
 import { SerializedNode, SerializedNodeWithId, NodeType, Attributes, IdNodeMap } from './types'
-import { getSerializedNodeId, hasSerializedNode, IGNORED_NODE_ID, setSerializedNode } from './serializationUtils'
+import {
+  absoluteToStylesheet,
+  getSerializedNodeId,
+  hasSerializedNode,
+  IGNORED_NODE_ID,
+  setSerializedNode,
+  transformAttribute,
+} from './serializationUtils'
 
 const tagNameRegex = /[^a-z1-6-_]/
 
@@ -41,108 +48,8 @@ function isCSSImportRule(rule: CSSRule): rule is CSSImportRule {
   return 'styleSheet' in rule
 }
 
-function extractOrigin(url: string): string {
-  let origin
-  if (url.indexOf('//') > -1) {
-    origin = url.split('/').slice(0, 3).join('/')
-  } else {
-    origin = url.split('/')[0]
-  }
-  origin = origin.split('?')[0]
-  return origin
-}
-
-const URL_IN_CSS_REF = /url\((?:(')([^']*)'|(")([^"]*)"|([^)]*))\)/gm
-const RELATIVE_PATH = /^(?!www\.|(?:http|ftp)s?:\/\/|[A-Za-z]:\\|\/\/).*/
-const DATA_URI = /^(data:)([^,]*),(.*)/i
-export function absoluteToStylesheet(cssText: string | null, href: string): string {
-  return (cssText || '').replace(
-    URL_IN_CSS_REF,
-    (origin: string, quote1: string, path1: string, quote2: string, path2: string, path3: string) => {
-      const filePath = path1 || path2 || path3
-      const maybeQuote = quote1 || quote2 || ''
-      if (!filePath) {
-        return origin
-      }
-      if (!RELATIVE_PATH.test(filePath)) {
-        return `url(${maybeQuote}${filePath}${maybeQuote})`
-      }
-      if (DATA_URI.test(filePath)) {
-        return `url(${maybeQuote}${filePath}${maybeQuote})`
-      }
-      if (filePath[0] === '/') {
-        return `url(${maybeQuote}${extractOrigin(href)}${filePath}${maybeQuote})`
-      }
-      const stack = href.split('/')
-      const parts = filePath.split('/')
-      stack.pop()
-      for (const part of parts) {
-        if (part === '.') {
-          continue
-        } else if (part === '..') {
-          stack.pop()
-        } else {
-          stack.push(part)
-        }
-      }
-      return `url(${maybeQuote}${stack.join('/')}${maybeQuote})`
-    }
-  )
-}
-
-function getAbsoluteSrcsetString(doc: Document, attributeValue: string) {
-  if (attributeValue.trim() === '') {
-    return attributeValue
-  }
-
-  const srcsetValues = attributeValue.split(',')
-  // srcset attributes is defined as such:
-  // srcset = "url size,url1 size1"
-  const resultingSrcsetString = srcsetValues
-    .map((srcItem) => {
-      // removing all but middle spaces
-      const trimmedSrcItem = srcItem.replace(/^\s+/, '').replace(/\s+$/, '')
-      const urlAndSize = trimmedSrcItem.split(' ')
-      // this means we have both 0:url and 1:size
-      if (urlAndSize.length === 2) {
-        const absUrl = absoluteToDoc(doc, urlAndSize[0])
-        return `${absUrl} ${urlAndSize[1]}`
-      } else if (urlAndSize.length === 1) {
-        const absUrl = absoluteToDoc(doc, urlAndSize[0])
-        return `${absUrl}`
-      }
-      return ''
-    })
-    .join(', ')
-
-  return resultingSrcsetString
-}
-
-export function absoluteToDoc(doc: Document, attributeValue: string): string {
-  if (!attributeValue || attributeValue.trim() === '') {
-    return attributeValue
-  }
-  const a: HTMLAnchorElement = doc.createElement('a')
-  a.href = attributeValue
-  return a.href
-}
-
 function isSVGElement(el: Element): boolean {
   return el.tagName === 'svg' || el instanceof SVGElement
-}
-
-export function transformAttribute(doc: Document, name: string, value: string): string {
-  // relative path in attribute
-  if (name === 'src' || (name === 'href' && value)) {
-    return absoluteToDoc(doc, value)
-  }
-  if (name === 'srcset' && value) {
-    return getAbsoluteSrcsetString(doc, value)
-  }
-  if (name === 'style' && value) {
-    return absoluteToStylesheet(value, location.href)
-  }
-  return value
 }
 
 function serializeNode(
